move auth code into datsources/spotify/auth. move typedefs and resolvers into separate files
