+++ src/app.js
@@ -7,10 +7,10 @@ const cors = require('cors');
 const SpotifyAPI = require('./datasources/spotify/SpotifyAPI');
 const MusixMatchAPI = require('./datasources/musixmatch/MusixMatchAPI');
 const routes = require('./routes/index');
-const {typeDefs, resolvers} = require('./schema');
-const {haveToken} = require('./schema/resolvers')
-const { ApolloServer } = require('apollo-server-express');
-//import schema from './schema/index';
+const {resolvers} = require('./schema/resolvers');
+const {typeDefs} = require('./schema/typeDefs');
+const {haveToken} = require('./datasources/spotify/auth/resolvers');
+const {ApolloServer} = require('apollo-server-express');
 
 const app = express();
 
@@ -20,8 +20,6 @@ const getDataSources = () => {
     musixMatchAPI: new MusixMatchAPI()
   };
 };
-
-haveToken().then(console.log);
 
 const server = new ApolloServer({
   typeDefs,
