+++ src/routes/index.js
@@ -8,39 +8,14 @@ router.get('/', function(req, res) {
 });
 
 
-const authorizeSpotify = require('../datasources/spotify/authorize');
-const getAccessToken = require('../datasources/spotify/getAccessToken');
-const getRecentlyPlayed = require('../datasources/spotify/getRecentlyPlayed')
+const authorizeSpotify = require('../datasources/spotify/auth/authorize');
+const getAccessToken = require('../datasources/spotify/auth/getAccessToken');
 
 router.get('/login', authorizeSpotify);
 router.get('/callback', getAccessToken, (req, res, next) => {
   //res.redirect(`${clientUrl}/?authorized=true`);
   res.redirect(`/graphql?access_token=${req.credentials.access_token}`); // todo use session
   console.log(`successfull authenticated`);
-});
-
-router.get('/history', (req, res) => {
-  db.find({}, (err, docs) => {
-    checkval = docs.length == 0;
-    console.log(`history, docsleng ${docs.length}, check ${checkval}`);
-    
-    if (err || docs.length == 0) {
-      res.redirect('/login');
-    }
-    else {
-    const accessToken = docs[0].access_token;
-    getRecentlyPlayed(accessToken)
-      .then(data => {
-        const arr = data.map(e => ({
-          played_at: e.played_at,
-          track_name: e.track.name,
-        }));
-
-        res.json(arr);
-      })
-      .catch(err => console.log(err));
-    }
-  });
 });
 
 module.exports = router;
